Update pico_main.py
Added LOTS of comments, hopefully the code is at least semi transparent. It works 110% of the time atm, if the proper command or improper command is given, it will always complete and loop back which is sick. 

There are block comments for where code can be put when a certain command is called. I assume most of the code should go on the pico for dead reckon functions to keep mom's code cleaner, but it's up to whoever writes it.

One thing to note. If there is a unknown function sent to the pico it will recognize it and reply "bad send". However Mom doesn't seem to recognize it and will report an error after 3 failed attempts. Atm not a huge deal but I do think this code is clean of errors currently.