Update mom_uart_commands.py
There are functions in this code that will be needed to copied to main.py. Send Command is the function that will be used if the pi wants to talk to pico.
wait for start will double check command was recieved.
wait for finish will read if command was completed IF wait for start worked.
Only Error I know of atm is that it does not read bad send correctly, but it does almost the same thing that it would if it did work atm.